Implement enemy aircraft move + stage processing

+++ Assets/Scripts/Object/IFollowingObject.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IFollowingObject
 {
-    void FollowTo(GameObject obj, double speed);
+    void FollowTo(GameObject obj, float speed);
 }
