+++ Assets/Scripts/Object/Projectiles/GuidedMissile.cs
@@ -5,7 +5,7 @@ public class GuidedMissile : Projectile, IFollowingObject
     public override ProjectileType ProjectileType => ProjectileType.GuidedMissile;
 
     
-    public void FollowTo(GameObject obj, double speed)
+    public void FollowTo(GameObject obj, float speed)
     {
         throw new System.NotImplementedException();
     }
